Вынесла валидацию в отдельный модуль
