+++ js/constant.js
@@ -5,8 +5,10 @@ const LIKE_MAX_COUNT = 200;
 const COMMENT_COUNT = 20;
 const COMMENTS_PORTION = 5;
 const ERROR_TAG_TEXT = 'В хэштегах допущены ошибки'; // сообщение об ошибке при написании хэштегов
+const COMMENTS_ERROR_MESSAGE = 'Максимальная длина комментария 140 символов';
 const VALID_SYMBOLS = /^#[a-zа-яё0-9]{1,19}$/i; // допустимые символы
 const MAX_HASHTAGS_COUNT = 5; // макс-ное кол-во хэштегов
+const MAX_COMMENTS_LENGTH = 140; // Максимальная длина комментария
 const COMMENT_LINES = [
   'Всё отлично!',
   'В целом всё неплохо.Но не всё.',
@@ -115,8 +117,10 @@ export {
   NAMES,
   COMMENTS_PORTION,
   ERROR_TAG_TEXT,
+  COMMENTS_ERROR_MESSAGE,
   VALID_SYMBOLS,
   MAX_HASHTAGS_COUNT,
+  MAX_COMMENTS_LENGTH,
   EFFECTS,
   STEP,
   MIN_VALUE,
