+++ js/form.js
@@ -1,5 +1,5 @@
 import { isEscapeKey, } from './util.js';
-import { ERROR_TAG_TEXT, VALID_SYMBOLS, MAX_HASHTAGS_COUNT } from './constant.js';
+import { pristineReset } from './validation.js';
 import { resetScale } from './scale.js';
 import { resetEffects } from './effects.js';
 
@@ -11,40 +11,6 @@ const body = document.querySelector('body');
 const fieldHashtags = uploadForm.querySelector('.text__hashtags');
 const fieldСomments = uploadForm.querySelector('.text__description');
 
-const SubmitButtonText = {
-  IDLE: 'Сохранить',
-  SENDING: 'Сохраняю...'
-};
-
-const pristine = new Pristine(uploadForm, {
-  classTo: 'img-upload__field-wrapper',
-  errorTextParent: 'img-upload__field-wrapper',
-  errorTextClass: 'img-upload__field-wrapper__error'
-});
-
-const blockSubmitButton = () => {
-  buttonCloseOverlay.disabled = true;
-  buttonCloseOverlay.textContent = SubmitButtonText.SENDING;
-};
-
-const unblockSubmitButton = () => {
-  buttonCloseOverlay.disabled = false;
-  buttonCloseOverlay.textContent = SubmitButtonText.IDLE;
-};
-
-const setFormSubmit = (cb) => {
-  uploadForm.addEventListener('submit', async (evt) => {
-    evt.preventDefault();
-
-    const isValid = pristine.validate();
-    if (isValid) {
-      blockSubmitButton();
-      await cb(new FormData(uploadForm));
-      unblockSubmitButton();
-    }
-  });
-};
-
 const onEscape = (evt) => {
   if (isEscapeKey(evt)) {
     evt.preventDefault();
@@ -52,37 +18,12 @@ const onEscape = (evt) => {
   }
 };
 
-// проверяет кол-во хэштегов
-const validHashtagCount = (tags) => tags.length <= MAX_HASHTAGS_COUNT;
-
-// проверяет уникальность хэштегов
-const uniqueHashtag = (tags) => {
-  const lowerCaseTags = tags.map((tag) => tag.toLowerCase());
-  return lowerCaseTags.length === new Set(lowerCaseTags).size;
-};
-
-// проверяем валидность
-const validHashtag = (tag) => VALID_SYMBOLS.test(tag);
-
-const validateTags = (value) => {
-  const tags = value
-    .trim()
-    .split(' ')
-    .filter((tag) => tag.trim().length);
-  return validHashtagCount(tags) && uniqueHashtag(tags) && tags.every(validHashtag);
-};
-
-pristine.addValidator(
-  fieldHashtags,
-  validateTags,
-  ERROR_TAG_TEXT
-);
-
 const fieldFocus = (field) => {
   field.addEventListener('focus', () => {
     document.removeEventListener('keydown', onEscape);
   });
 };
+
 const fieldBlur = (field) => {
   field.addEventListener('blur', () => {
     document.addEventListener('keydown', onEscape);
@@ -118,7 +59,7 @@ function closeImageModal() {
   body.classList.remove('modal-open');
   document.removeEventListener('keydown', onEscape);
   uploadForm.reset();
-  pristine.reset();
+  pristineReset();
   focusRemove();
   resetScale();
   resetEffects();
@@ -136,5 +77,6 @@ const editImages = () => {
 };
 
 uploadFile.addEventListener('input', editImages);
-export { setFormSubmit, closeImageModal, onEscape, uploadFile };
+
+export { closeImageModal, onEscape, uploadFile };
